+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 
 Check [Keep a Changelog](https://keepachangelog.com/) for recommendations on how to structure this file.
+
+
+## 5.0.0
+- display format error on problems panel [#37](https://github.com/foxundermoon/vs-shell-format/issues/37)
+- remove showError configuration
+
 ## 4.0.11
 - resolve 1 vulnerability  https://www.npmjs.com/advisories/803